Fixed seed bug

+++ server/database/seed.js
@@ -1,15 +1,15 @@
 const { User, Item } = require('./models');
 const { users, items } = require('./data');
-const mongoose  = require('mongoose');
 const seed = async() => {
     console.log("Seeding...")
-
-    User.collection.drop();
+    const usersCount = await User.collection.countDocuments({});
+    if( usersCount !== 0) User.collection.drop();
     const Users = [];
     for (let u of users) 
         Users.push(User.create(u));
 
-    Item.collection.drop();
+    const itemsCount = await Item.collection.countDocuments({});
+    if(itemsCount !== 0) Item.collection.drop();
     const Items = [];
     for (let i of items)
         Items.push(Item.create(i));
